+++ backend/src/config/prisma.ts
@@ -1,4 +1,9 @@
-import { PrismaClient } from "@prisma/client";
+import { PrismaClient } from "../prisma/generated/prisma/client.js";
+import { PrismaPg } from "@prisma/adapter-pg";
+import { envConfig } from "./env.js";
 
-const prisma = new PrismaClient();
-export default prisma;
+const adapter = new PrismaPg({
+  connectionString: envConfig.databaseUrl
+});
+
+export const prisma = new PrismaClient({ adapter });